Separate EditableFeatureCollection from ImmutableFeatureCollection
