+++ modules/core/src/test/lib/editable-feature-collection.test.js
@@ -170,428 +170,6 @@ describe('setDrawAtFront()', () => {
   });
 });
 
-describe('replacePosition()', () => {
-  it(`doesn't mutate original`, () => {
-    const leaveMeAlone = {
-      type: 'FeatureCollection',
-      features: [pointFeature]
-    };
-    const features = new EditableFeatureCollection(leaveMeAlone);
-    features.replacePosition(0, [], [10, 20]);
-
-    expect(leaveMeAlone.features[0].geometry.coordinates).toEqual([1, 2]);
-  });
-
-  it('replaces position in Point', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [pointFeature]
-    });
-    features.replacePosition(0, [], [10, 20]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [10, 20];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('replaces first position in LineString', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    });
-    features.replacePosition(0, [0], [10, 20]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [[10, 20], [2, 3], [3, 4]];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('replaces middle position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.replacePosition(0, [0, 1], [1.1, -1.1]).replacePosition(0, [1, 2], [0.6, 0.6]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1, -1], [1.1, -1.1], [1, 1], [-1, 1], [-1, -1]],
-      [[-0.5, -0.5], [-0.5, 0.5], [0.6, 0.6], [0.5, -0.5], [-0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('replaces last position when replacing first position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.replacePosition(0, [0, 0], [-1.1, -1.1]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1.1, -1.1], [1, -1], [1, 1], [-1, 1], [-1.1, -1.1]],
-      [[-0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [0.5, -0.5], [-0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('replaces first position when replacing last position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.replacePosition(0, [0, 4], [-1.1, -1.1]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1.1, -1.1], [1, -1], [1, 1], [-1, 1], [-1.1, -1.1]],
-      [[-0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [0.5, -0.5], [-0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-});
-
-describe('removePosition()', () => {
-  it(`doesn't mutate original`, () => {
-    const leaveMeAlone = {
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    };
-    const features = new EditableFeatureCollection(leaveMeAlone);
-    features.removePosition(0, [0]);
-
-    expect(leaveMeAlone.features[0].geometry.coordinates).toEqual([[1, 2], [2, 3], [3, 4]]);
-  });
-
-  it('throws exception when attempting to remove Point', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [pointFeature]
-    });
-
-    expect(() => features.removePosition(0, [0])).toThrow(
-      `Can't remove a position from a Point or there'd be nothing left`
-    );
-  });
-
-  it('removes first position in LineString', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    });
-    features.removePosition(0, [0]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [[2, 3], [3, 4]];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('removes middle position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.removePosition(0, [0, 1]).removePosition(0, [1, 3]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1, -1], [1, 1], [-1, 1], [-1, -1]],
-      [[-0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [-0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('changes last position when removing first position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.removePosition(0, [1, 0]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1, -1], [1, -1], [1, 1], [-1, 1], [-1, -1]],
-      [[-0.5, 0.5], [0.5, 0.5], [0.5, -0.5], [-0.5, 0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('changes first position when removing last position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.removePosition(0, [1, 4]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1, -1], [1, -1], [1, 1], [-1, 1], [-1, -1]],
-      [[0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('throws exception when LineString has only 2 positions', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [
-        { type: 'Feature', geometry: { type: 'LineString', coordinates: [[0, 1], [2, 3]] } }
-      ]
-    });
-
-    expect(() => features.removePosition(0, [0])).toThrow(
-      `Can't remove position. LineString must have at least two positions`
-    );
-  });
-
-  it('throws exception when Polygon outer ring has only 4 positions (triangle)', () => {
-    let features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features = features
-      // Convert from quadrilateral to triangle
-      .removePosition(0, [0, 1]);
-
-    expect(() => features.removePosition(0, [0, 1])).toThrow(
-      `Can't remove position. Polygon's outer ring must have at least four positions`
-    );
-  });
-
-  it('removes hole from Polygon when it has less than four positions', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.removePosition(0, [1, 1]).removePosition(0, [1, 1]);
-
-    const actualGeometry = features.getFeatureCollection().features[0].geometry;
-    const expectedGeometry = {
-      type: 'Polygon',
-      coordinates: [[[-1, -1], [1, -1], [1, 1], [-1, 1], [-1, -1]]]
-    };
-
-    expect(actualGeometry).toEqual(expectedGeometry);
-  });
-
-  it('removes LineString from MultiLineString when it has only one position', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiLineStringFeature]
-    });
-    features.removePosition(0, [1, 0]).removePosition(0, [1, 0]);
-
-    const actualGeometry = features.getFeatureCollection().features[0].geometry;
-    const expectedGeometry = {
-      type: 'MultiLineString',
-      coordinates: [[[1, 2], [2, 3], [3, 4]]]
-    };
-
-    expect(actualGeometry).toEqual(expectedGeometry);
-  });
-
-  it('throws exception when MultiLineString has only 2 positions', () => {
-    let features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiLineStringFeature]
-    });
-    features = features
-      // remove one of the LineStrings
-      .removePosition(0, [0, 0])
-      .removePosition(0, [0, 0])
-      // shrink the remaining LineString to two positions
-      .removePosition(0, [0, 0]);
-
-    expect(() => features.removePosition(0, [0, 0])).toThrow(
-      `Can't remove position. MultiLineString must have at least two positions`
-    );
-  });
-
-  it('removes Polygon from MultiPolygon when outer ring has less than four positions', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiPolygonFeature]
-    });
-    features.removePosition(0, [1, 0, 0]).removePosition(0, [1, 0, 0]);
-
-    const actualGeometry = features.getFeatureCollection().features[0].geometry;
-    const expectedGeometry = {
-      type: 'MultiPolygon',
-      coordinates: [multiPolygonFeature.geometry.coordinates[0]]
-    };
-
-    expect(actualGeometry).toEqual(expectedGeometry);
-  });
-
-  it('removes hole from MultiPolygon when it has less than four positions', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiPolygonFeature]
-    });
-    features.removePosition(0, [0, 1, 0]).removePosition(0, [0, 1, 0]);
-
-    const actualGeometry = features.getFeatureCollection().features[0].geometry;
-    const expectedGeometry = {
-      type: 'MultiPolygon',
-      coordinates: [
-        [[[-1, -1], [1, -1], [1, 1], [-1, 1], [-1, -1]]],
-        [[[2, -1], [4, -1], [4, 1], [2, 1], [2, -1]]]
-      ]
-    };
-
-    expect(actualGeometry).toEqual(expectedGeometry);
-  });
-
-  it('throws exception when MultiPolygon outer ring has only 4 positions', () => {
-    let features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiPolygonFeature]
-    });
-    features = features
-      // Remove the second polygon
-      .removePosition(0, [1, 0, 0])
-      .removePosition(0, [1, 0, 0])
-      // Remove positions from outer ring
-      .removePosition(0, [0, 0, 1]);
-
-    expect(() => features.removePosition(0, [0, 0, 1])).toThrow(
-      `Can't remove position. MultiPolygon's outer ring must have at least four positions`
-    );
-  });
-});
-
-describe('addPosition()', () => {
-  it(`doesn't mutate original`, () => {
-    const leaveMeAlone = {
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    };
-    const features = new EditableFeatureCollection(leaveMeAlone);
-    features.addPosition(0, [1], [2, 3]);
-
-    expect(leaveMeAlone.features[0].geometry.coordinates).toEqual([[1, 2], [2, 3], [3, 4]]);
-  });
-
-  it('throws exception when attempting to add position to Point', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [pointFeature]
-    });
-
-    expect(() => features.addPosition(0, [], [3, 4])).toThrow(
-      'Unable to add a position to a Point feature'
-    );
-  });
-
-  it('adds position to beginning of LineString', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    });
-    features.addPosition(0, [0], [10, 20]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [[10, 20], [1, 2], [2, 3], [3, 4]];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('adds position to middle of LineString', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    });
-    features.addPosition(0, [1], [10, 20]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [[1, 2], [10, 20], [2, 3], [3, 4]];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('adds position to end of LineString', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [lineStringFeature]
-    });
-    features.addPosition(0, [3], [10, 20]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [[1, 2], [2, 3], [3, 4], [10, 20]];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-
-  it('adds position in Polygon', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [polygonFeature]
-    });
-    features.addPosition(0, [0, 1], [0, -1]).addPosition(0, [1, 4], [0, -0.5]);
-
-    const actualCoordinates = features.getFeatureCollection().features[0].geometry.coordinates;
-    const expectedCoordinates = [
-      [[-1, -1], [0, -1], [1, -1], [1, 1], [-1, 1], [-1, -1]],
-      [[-0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [0.5, -0.5], [0, -0.5], [-0.5, -0.5]]
-    ];
-
-    expect(actualCoordinates).toEqual(expectedCoordinates);
-  });
-});
-
-describe('addFeature()', () => {
-  it(`doesn't mutate original`, () => {
-    const leaveMeAlone = {
-      type: 'FeatureCollection',
-      features: []
-    };
-    const features = new EditableFeatureCollection(leaveMeAlone);
-    features.addFeature(pointFeature);
-
-    expect(leaveMeAlone.features.length).toEqual(0);
-  });
-
-  it('adds feature to empty array', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: []
-    });
-    const actualFeatures = features.addFeature(pointFeature).getFeatureCollection();
-
-    const expectedFeatures = {
-      type: 'FeatureCollection',
-      features: [pointFeature]
-    };
-
-    expect(actualFeatures).toEqual(expectedFeatures);
-  });
-
-  it('adds feature to end of array', () => {
-    const features = new EditableFeatureCollection({
-      type: 'FeatureCollection',
-      features: [multiPointFeature]
-    });
-    const actualFeatures = features.addFeature(multiLineStringFeature).getFeatureCollection();
-
-    const expectedFeatures = {
-      type: 'FeatureCollection',
-      features: [multiPointFeature, multiLineStringFeature]
-    };
-
-    expect(actualFeatures).toEqual(expectedFeatures);
-  });
-});
-
 describe('getEditHandles()', () => {
   it('gets edit handles for Point', () => {
     const features = new EditableFeatureCollection({
@@ -815,7 +393,7 @@ describe('drawLineString mode', () => {
             }
           ]
         },
-        featureIndex: featureCollection.features.length + 1,
+        featureIndex: featureCollection.features.length,
         position: null,
         positionIndexes: null
       });
